+++ R/jumpTests.R
@@ -95,7 +95,8 @@ ABDJumptest <- function(RV, BPV, TQ) { # Compute jump detection stat mentioned i
 #' 
 #' @author Giang Nguyen, Jonathan Cornelissen, Kris Boudt, and Emil Sjoerup.
 #'
-#' @examples
+#' @examplesIf !grepl("debian", sessionInfo()["platform"], fixed = TRUE)
+#' \dontshow{data.table::setDTthreads(2)}
 #' jt <- AJjumpTest(sampleTData[, list(DT, PRICE)], p = 2, k = 3, 
 #'                  alignBy = "seconds", alignPeriod = 5, makeReturns = TRUE)
 #'